Hoist is_confirmed_rooted() to its only caller
BlockCommitmentCache no longer depends on Blockstore
